refactor: 수량 관련 로직 커스텀 훅 분리
